Added PlayerDirection class
This class can be used to determine what 'direction' the player is
facing. This can be used to allow the player to interact with the item
or the door that they are facing. NOTE: I may have to refactor this and
change the name to Direction to allow different items to have a
direction too for drawing. But for the initial design and basic
implementation this should be enough to allow you to begin drawing UI