Test Class for testing features.
Running the main method of this class will allow you to test and see the
features of the current model. The console will print out the room.
Movement of the player can be tested and seen inside the room, picking
up items can be viewed and also dropping item. Note dropping item finds
the adjacent empty space.