Added 2d array of Pieces to Room class..
Now a Room instance contains a 2D array of Pieces that represent the
internal structure of the Room. Piece, such as Player or items, can be
added to specific locations in the 2D array. 
To allow pieces to be added and removed, respective methods have been
included.
Also, printRoom() method has been added to allow a room to be printed on
the console for testing.