Updated the Test class to reflect changes
Running the main method of this class will allow you to test and see the
features of the current model. The console will print out the room.
Movement of the player can be tested and seen inside the room, picking
up items can be viewed and also dropping item. 

There are now two rooms and a player has to pickup the key before going
to the other room