Added Byte Array functionality
The state of the pieces on the board can now be converted to and loaded
from a byte array. This will allow the client class to receive updates
to the game state. The game can also be saved by converting it to a byte
array to allow the game to restart.