Added byte array functionality for door
The door is a special piece as it links two separate two rooms together.
The inputStream and outputStream methods allow the door state to be
converted into a byte array for allowing the client's to load the game.
The door piece is the last piece that gets converted into a byte array
as it requires all the rooms to be created first. For this reason there
is an extra getRoom() method that retrieves the room that the door leads
to.