Added byte array functionality for Room
The room contains a 2d array of pieces which represents the items inside
the room. The inputStream and outputStream methods allow the room state
to be converted into a byte array for allowing the client's to load the
game. The outputStream and inputStream methods call the piece
outputStream and inputStream methods respectively to save/load the state
of each player.