Allow game to convert to a byte array
Each room in the game is converted to a byte array by calling its
outputStream method, which then calls the outputStream method of all the
pieces that are inside the room. The byte array is written to the
DataOutputStream which is then returned.
For loading the game from the byte array, each room's inputStream method
is called which then calls the inputStream method of all the pieces that
are to be placed inside the room.
This allows the game to completely load from a byte array.