Updated the Test class to reflect changes
Running the main method of this class will allow you to test and see the
features of the current model. Currently, the movement of the player can
be tested and the transition of a player from one room to another works.
This door is locked and the player must get the key before it can go
past the door.