Added Exceptions and Refactored Control class
InvalidMove exception has been added to the control class that allows
the InvalidMove exception to be thrown when a user attempts to make an
invalid move such as moving on top of a player or attempting to drop
items from an already empty player container. These exceptions are used
to test the program through JUnit tests and also allow the game to catch
and differentiate between errors and perform appropriate action.

The class has been refactored and ready for the final merge.