Added and refactored Parser class
This class allows the Game to be initialised from the Living.txt file.
This file needs to be stored in the src folder and contains the number
of rooms and the initial state of all the pieces in each room. The
getGameFromFile needs to be called to add the rooms and pieces in game. 
The Parser class has been refactored and is ready to be merged with the
master.