Added equals and hashCode and refactored class
equals and hashCode methods have been added, this is mainly for junit
tests as they involve comparing positions. The position objects are not
compared or checked for equality in the rest of the program so they
should not cause any unexpected behaviour. 
Javadocs and comments have been added to ensure that this class is
readable.
The Position class is ready for the final merge.