Added equals and hashCode and refactored class
equals and hashCode methods have been added, this is mainly for junit
tests as they involve comparing differnent location objects. The
location objects are not compared or checked for equality in the rest of
the program so they should not cause any unexpected behaviour. 
Javadocs and comments have been added to ensure that this class is
readable.
The Location class is ready for the final merge.