Updated the Test class to reflect changes
Running the main method of this class will allow you to test and see the
features of the current model. Currently, the movement of the player can
be tested and the transition of a player from one room to another works.
Drop item functionality has been changed. Now dropping item involves
taking the last added item and removing it. Essentially the drop
functionality causes the player's container to resemble a stack.